+++ proof-node/__main__.py
@@ -1,8 +1,10 @@
+import random
+import time
 import logging
 from flask import Flask, request, jsonify
 import docker
-import tempfile
 import os
+import tempfile
 
 logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(levelname)s - %(message)s')
 logger = logging.getLogger(__name__)
@@ -24,7 +26,7 @@ fs.mounts = [
   {{ path = "/usr", uri = "file:/usr" }},
   {{ path = "/etc", uri = "file:/etc" }},
   {{ path = "/bin", uri = "file:/bin" }},
-  {{ path = "/app", uri = "file:/app" }},
+  {{ path = "/root", uri = "file:/root" }},
 ]
 
 sgx.debug = true
@@ -38,18 +40,19 @@ sgx.trusted_files = [
   "file:/usr/lib/python3.9/",
   "file:/usr/local/lib/python3.9/",
   "file:/lib/x86_64-linux-gnu/",
-  "file:/app/print_env.py",
 ]
+
+loader.env.PYTHON_SCRIPT = "{env_vars.get('PYTHON_SCRIPT', 'print("No script provided")')}"
 """
 
     for key, value in env_vars.items():
-        manifest_content += f'loader.env.{key} = "{value}"
-'
+        if key != 'PYTHON_SCRIPT':
+            manifest_content += f'loader.env.{key} = "{value}"\n'
 
     return manifest_content
 
-@app.route('/run', methods=['POST'])
-def run_container():
+@app.route('/spunup', methods=['POST'])
+def spunup():
     data = request.json
     if not data or 'image_url' not in data or 'env_vars' not in data:
         return jsonify({"error": "Missing image_url or env_vars"}), 400
@@ -69,13 +72,14 @@ def run_container():
             with open(manifest_path, 'w') as f:
                 f.write(manifest_content)
 
-            # Run the container with Gramine-SGX
+            # Run the container with Gramine
             container = docker_client.containers.run(
                 image_url,
                 command=[
                     "gramine-sgx",
                     "python",
-                    "/app/print_env.py"
+                    "-c",
+                    env_vars.get('PYTHON_SCRIPT', 'print("No script provided")')
                 ],
                 volumes={
                     manifest_path: {'bind': '/python.manifest', 'mode': 'ro'},
@@ -83,14 +87,15 @@ def run_container():
                 },
                 devices=['/dev/sgx_enclave'],
                 environment=env_vars,
+                detach=True,
                 remove=True
             )
 
-            # Get logs
-            logs = container.logs().decode()
-            logger.info(f"Container logs: {logs}")
+            # Wait for the container to finish and get logs
+            logs = container.logs(stream=True)
+            result = "".join([log.decode() for log in logs])
 
-            return jsonify({"result": logs})
+            return jsonify({"result": result, "exit_code": container.wait()['StatusCode']})
     except docker.errors.DockerException as e:
         logger.error(f"Docker error: {str(e)}")
         return jsonify({"error": str(e)}), 500
@@ -99,4 +104,5 @@ def run_container():
         return jsonify({"error": "An unexpected error occurred"}), 500
 
 if __name__ == "__main__":
+    logging.info("Starting main")
     app.run(host='0.0.0.0', port=5000)
